splits AlgoliaController and deletes PreviewContent
