+++ Sup_Test_iOS_InstantSearch/StockItem.swift
@@ -1,0 +1,8 @@
+//
+//  StockItem.swift
+//  Sup_Test_iOS_InstantSearch
+//
+//  Created by Steven Kaiser on 8/11/23.
+//
+
+import Foundation
